Refactor of main page
