+++ public/index.html
@@ -13,13 +13,11 @@
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Libre+Baskerville:wght@700&display=swap" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap" rel="stylesheet">
 
-    <link href="https://fonts.googleapis.com/css2?family=IBM+Plex+Sans:wght@500&family=IBM+Plex+Serif:wght@300;400&display=swap" rel="stylesheet">
+    
 
-    <link href="https://fonts.googleapis.com/css2?family=Koulen&family=Staatliches&display=swap" rel="stylesheet">
-
-    <title>CineTEX</title>
+    <title>Cinex Manager</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
