+++ src/components/SearchMovieForm.js
@@ -1,5 +1,5 @@
 import '../styles/components.css';
-import NavBar from './NavBar';
+import NavBar from './navbar/navbar';
 import React, { useState } from 'react';
 import Modal from 'react-modal';
 import {useTable} from "react-table";
