+++ src/components/main_screen/main_screen.css
@@ -1,0 +1,64 @@
+.main_container {
+    width: 100%;
+    height: 100vh;
+    display: grid;
+    grid-template-columns: 50% 50%;
+    text-align: center;
+
+}
+
+._title_container {
+    align-items: center;
+    margin-top: 5%;
+    margin-left: 15%;
+}
+
+h1 {
+    font-family: "Roboto", sans-serif;
+    font-weight: 800;
+    font-size: 130px;
+    color: white;
+    text-align: start;
+    text-shadow: 
+        1px 0px 1px #444343,
+        3px 0px 0px #444343,
+        5px 0px 0px #444343,
+        5px 0px 0px #444343,
+        5px 0px 0px #444343,
+        5px 0px 0px #444343,
+        5px 0px 0px #444343,
+        5px 0px 0px #444343,
+        5px 0px 0px #444343,
+        5px 0px 0px #444343,
+        7px 0px 0px rgba(0, 0, 0, 0.5);
+}
+
+._options_container {
+    position: relative;
+    background-color: #252525;
+    width: 400px;
+    height: 400px;
+    padding: 10px;
+    margin-top: 17%;
+    margin-left: 20%;
+    border-radius: 3px;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+
+}
+
+.button_search_movie, .button_add_movie {
+    font-family: "Roboto", sans-serif;
+    margin-top: 5%;
+    position: relative;
+    width: 300px;
+    height: 70px;
+    border: 0px;
+    font-weight: 500;
+    color: rgb(255, 255, 255);
+    font-size: 20px;
+    background-color:rgb(71, 66, 66);
+    border-radius: 3px;
+    cursor:pointer;
+}
