+++ src/components/navbar/navbar.css
@@ -1,0 +1,35 @@
+nav {
+    width: 100%;
+    height: 100px;
+    background-color: transparent;
+}
+
+.button_home {
+    background-color: transparent;
+    color: white;
+    padding-top: 50px;
+    position: relative;
+
+    left: 7%;
+    border: 0;
+    font-family: "Roboto", sans-serif;
+    font-style: normal;
+    font-size: 20px;
+    cursor: pointer;
+}
+
+img {
+    position: relative;
+    padding: 10px;
+}
+
+.logo {
+    position: relative;
+    background-image: url(/public/images/logo.png);
+    background-position: center;
+    width:300px;
+    height: 70px;
+    padding: 10px;
+    left: 2%;
+    
+}
